requirements.txt
streamlit
PyMuPDF
pytesseract
Pillow
deepl
kraken

